% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADH_global}
\alias{ADH_global}
\title{ADH China Shock: "Global" Industry Growths}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 780 rows and 3 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
ADH_global
}
\description{
Contains industry growth at the year-industry level.
}
\keyword{datasets}
