% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADH_local}
\alias{ADH_local}
\title{ADH China Shock: Local Industry Shares}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 563160 rows and 4 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
ADH_local
}
\description{
Contains local industry share at the community-year-industry level.
}
\keyword{datasets}
