% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADH_local_wide}
\alias{ADH_local_wide}
\title{ADH China Shock: Local Industry Shares in Wide Format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1444 rows and 782 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
ADH_local_wide
}
\description{
Contains local industry share at the community-year level.
}
\keyword{datasets}
