% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADH_master}
\alias{ADH_master}
\title{ADH China Shock: Master Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1444 rows and 20 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
ADH_master
}
\description{
Contains main variables at the community-year level.
}
\keyword{datasets}
