% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAR_local}
\alias{BAR_local}
\title{BAR Canonical Bartik: Local Industry Shares}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 493848 rows and 5 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
BAR_local
}
\description{
Contains local industry share at the community-year-industry level.
}
\keyword{datasets}
