% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAR_local_wide}
\alias{BAR_local_wide}
\title{BAR Canonical Bartik: Local Industry Shares in Wide Format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2166 rows and 686 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
BAR_local_wide
}
\description{
Contains local industry share at the community-year level.
}
\keyword{datasets}
