% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAR_master}
\alias{BAR_master}
\title{BAR Canonical Bartik: Master Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2166 rows and 742 columns.}
\source{
\url{https://github.com/paulgp/bartik-weight}
}
\usage{
BAR_master
}
\description{
Contains main variables at the community-year level.
}
\keyword{datasets}
