% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartik.weight-package.r
\docType{package}
\name{bartik.weight}
\alias{bartik.weight}
\title{bartik.weight.}
\description{
Contains example datasets and a function for estimating Rotemberg
weights outlined in Goldsmith-Pinkham, Sorkin and Swift (2018).
}
