% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\title{Estimating Rotemberg Weights for a Bartik "Instrument"}
\usage{
bw(
  master,
  y,
  x,
  controls = NULL,
  weight = NULL,
  local,
  z,
  global,
  g,
  L_id,
  T_id = NULL
)
}
\arguments{
\item{master}{The master data frame.}

\item{y}{A string for outcome variable. It should be a variable in \code{master}.}

\item{x}{A string for the endogenous variable. It should be a variable in
\code{master}.}

\item{controls}{A string or character vector for the control variables.
\code{controls} are optional and should be in \code{master}.}

\item{weight}{A string for the weighted variable. \code{weight} is optional and
should be in \code{master}.}

\item{local}{The local data frame. It should be in wide format.}

\item{z}{The variables used in the local data frame.}

\item{global}{The global data frame.}

\item{g}{The variables used in the global data frame.}

\item{L_id}{The indexes used to identify and match the locations in the
master and local data frames.}

\item{T_id}{The indexes used to identify and match the periods in the master
and global data frames.}
}
\value{
A tibble with K rows, a for each group, with :
\itemize{
\item the \code{global} dataset.
\item one column \code{alpha} for the Rotemberg weights.
\item one column \code{beta} for the just-identified IV estimated.
\item one column \code{gamma} reduced form/intent-to-treat effect of the Bartik IV on Y.
\item one column \code{pi} the first stage effect of the Bartik IV on X.
}
}
\description{
\code{bw()} estimates the Rotemberg weights for a Bartik "instrument" outlined in
Goldsmith-Pinkham, Sorkin and Swift (2018) and the just-identified IV
estimates using local industry shares, which are the actual instruments.
}
\details{
The typical Bartik "instrument" has two components: the local industry
shares (usually at the level of location-industry or location-year-industry),
and the overall industry growth rates (usually at the level of industry or
industry-year). The outcome variable and the endogenous variable are at the
level of location or location-year.

Regarding the application in the economics of migration, the shock comes
from the changes in the population of migrants, the shift. The share is the
distribution of migrants, grouped by origin, across the locations in a
previous period of reference.

Because the key variables are at three different levels, \code{bw()} proceeds by
taking three different datasets: (1) a main \code{master} data frame containing
the dependent variable (\code{y}), the causal variable of interest (\code{x}), a set of
control variables (\code{controls}), and the weighted variables (\code{weight}); (2) a
\code{local} data frame containing the shares (\code{Z}); and (3) a \code{global} data frame
containing the shift (\code{G}). At the moment, it is necessary to transform the
"local" dataset from long format to wide format.

Given the dimensions:
\itemize{
\item N the number of observations
\item L the number of locations
\item K the number of groups
\item T the number of periods
}

The dimensions of these datasets are :
\itemize{
\item \code{master} : N rows
\item \code{local[Z]} : L rows and K columns
\item \code{global[G]} : K rows and T columns
}
}
\examples{
library(bartik.weight)

index = c("czone", "year")
y = "d_sh_empl_mfg"
x = "d_tradeusch_pw"
controls = c("reg_midatl", "reg_encen", "reg_wncen", "reg_satl",
  "reg_escen", "reg_wscen", "reg_mount", "reg_pacif", "l_sh_popedu_c",
  "l_sh_popfborn", "l_sh_empl_f", "l_sh_routine33", "l_task_outsource",
  "t2", "l_shind_manuf_cbp")
weight = "timepwt48"
Z = setdiff(names(ADH_local_wide), index)
G = "trade_"

bw(ADH_master, y, x, controls, weight,
   ADH_local_wide, Z, ADH_global, G,
   L_id = index, T_id = NULL)
}
