% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\title{Estimating Rotemberg Weights for a Bartik "Instrument"}
\usage{
bw(master, y, x, controls = NULL, weight = NULL, local, Z, global, G)
}
\arguments{
\item{master}{The master data frame.}

\item{y}{A string for outcome variable. It should be a variable in \code{master}.}

\item{x}{A string for the endogenous variable. It should be a variable in
\code{master}.}

\item{controls}{A string or character vector for the control variables.
\code{controls} are optional and should be in \code{master}.}

\item{weight}{A string for the weighted variable. \code{weight} is optional and
should be in \code{master}.}

\item{local}{The local data frame. It should be in wide format.}

\item{Z}{A string or character vector for the the local industry shares.}

\item{global}{The global data frame.}

\item{G}{A string for the the overall industry growth rates.}
}
\value{
A data frame built from \code{global} with two additional variables:
\code{alpha} (the Rotemberg weights), and \code{beta} (the just-identified IV
estimates).
}
\description{
\code{bw()} estimates the Rotemberg weights for a Bartik "instrument" outlined in
Goldsmith-Pinkham, Sorkin and Swift (2018) and the just-identified IV
estimates using local industry shares, which are the actual instruments.
}
\details{
The typical Bartik "instrument" has two components: the local industry
shares (usually at the level of location-industry or location-year-industry),
and the overall industry growth rates (usually at the level of industry or
industry-year). The outcome variable and the endogenous variable are at the
level of location or location-year.

Because the key variables are at three different levels, \code{bw()} proceeds by
taking three different datasets: (1) a main \code{master} data frame containing
the dependent variable (\code{y}), the causal variable of interest (\code{x}), a set of
control variables (\code{controls}), and the weighted variables (\code{weight}); (2) a
\code{local} data frame containing the set of local industry shares (\code{Z}); and (3)
a \code{global} data frame containing the overall industry growth rates (\code{G}). At
the moment, it is necessary to transform the "local" dataset from long format
to wide format.
}
\examples{
library(bartik.weight)

index = c("czone", "year")
y = "d_sh_empl_mfg"
x = "d_tradeusch_pw"
controls = c("reg_midatl", "reg_encen", "reg_wncen", "reg_satl",
  "reg_escen", "reg_wscen", "reg_mount", "reg_pacif", "l_sh_popedu_c",
  "l_sh_popfborn", "l_sh_empl_f", "l_sh_routine33", "l_task_outsource",
  "t2", "l_shind_manuf_cbp")
weight = "timepwt48"
Z = setdiff(names(ADH_local_wide), index)
G = "trade_"

bw(ADH_master, y, x, controls, weight, ADH_local_wide, Z, ADH_global, G)

}
